requirements.txt
faiss-cpu
sentence-transformers
requests
streamlit
numpy
scikit-learn
pypdf
